Ticketing Examples IO: Update code
