+++ src/test/resources/test-data/ticketing/output-samples/customer-sample.json
