+++ src/test/resources/test-data/ticketing/output-samples/event-sample.json
