+++ src/test/resources/test-data/ticketing/output-samples/offer-sample.json
