+++ src/test/resources/test-data/ticketing/output-samples/purchaseCompleted-sample.json
