+++ src/test/resources/test-data/ticketing/output-samples/seat-sample.json
