+++ src/test/resources/test-data/ticketing/output-samples/seatAssigned-sample.json
