+++ src/test/resources/test-data/ticketing/output-samples/seatUnassigned-sample.json
