+++ src/test/resources/test-data/ticketing/output-samples/ticket-sample.json
