+++ src/test/resources/test-data/ticketing/output-samples/ticketUsed-sample.json
