+++ src/test/resources/test-data/ticketing/output-samples/ticketUsedIO-sample.json
@@ -1,0 +1,1 @@
+{"occurred_at":"2017-12-13T10:18:27.496Z","type":"as.event.ticket.scan.entered","origin":"spacejam.scanning","involves":[{"ACTOR":{"entity_ref":"Customer/951"}},{"INVOLVES":{"entity_ref":"Venue/7856"}},{"INVOLVES":{"entity_ref":"Gate/055"}},{"INVOLVES":{"entity_ref":"Ticket/852456"}},{"AFFECTS:GRANTS_ACCESS_TO":{"entity_ref":"Seat/23_2_12"}}],"importance":3,"properties":{"scanned":"true"}}
